+++ graalpython/com.oracle.graal.python/src/com/oracle/graal/python/nodes/SpecialMethodNames.java
@@ -161,6 +161,7 @@ public abstract class SpecialMethodNames {
     public static final String TOBYTES = "tobytes";
     public static final String DECODE = "decode";
     public static final String __SIZEOF__ = "__sizeof__";
+    public static final String __CLASS_GETITEM__ = "__class_getitem__";
 
     public static final String RICHCMP = "__truffle_richcompare__";
     public static final String TRUFFLE_SOURCE = "__truffle_source__";